fix(parquet): Prevent int32 overflow when writing large dictionary pages
Writing a dictionary-encoded column whose page holds more than ~2 GB of encoded indices produced a corrupt Parquet file. The buffer-size estimates for the RLE/bit-packed index stream were computed in `int`, so the byte count wrapped around and the writer under-allocated. This ports the upstream fix (apache/arrow GH-47973): the size math in `RleEncoder`, `LevelEncoder`, the dictionary encoder, and `BitWriter`'s bounds check now uses `int64`, and the encoders throw when a page would exceed the `int32` byte limit instead of silently truncating.

The upstream reproducer writes ~150 GB, so the regression test instead checks `RleEncoder::MaxBufferSize` directly: for a wide bit width and a large value count it must return a size above `int32` max, which overflowed before the fix.

+++ velox/dwio/parquet/common/BitStreamUtilsInternal.h
@@ -224,7 +224,9 @@ inline bool BitWriter::PutValue(uint64_t v, int numBits) {
     VELOX_DCHECK_EQ(v >> numBits, 0, "v = {}, numBits = {}", v, numBits);
   }
 
-  if (FOLLY_UNLIKELY(byteOffset_ * 8 + bitOffset_ + numBits > maxBytes_ * 8))
+  if (FOLLY_UNLIKELY(
+          static_cast<int64_t>(byteOffset_) * 8 + bitOffset_ + numBits >
+          static_cast<int64_t>(maxBytes_) * 8))
     return false;
 
   bufferedValues_ |= v << bitOffset_;
